# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Meu Auto API
  version: "1.0.0"
  description: |
    API do Meu Auto — acompanhamento da vida útil de um veículo para o dono pessoa física
    no Brasil.

    **Convenções**

    - Datas civis (dia de uma manutenção, vencimento de IPVA) são `YYYY-MM-DD`, sem hora e
      sem fuso. Timestamps de auditoria (`created_at`, `updated_at`) são RFC 3339.
    - Dinheiro é sempre **inteiro em centavos** (`*_cents`). Nunca float, nunca string.
    - Distância é sempre inteiro em quilômetros.
    - Mensagens de erro são pt-BR e podem ser exibidas ao usuário. O `code` é estável e em
      inglês — é nele que o cliente decide comportamento, nunca na mensagem.
    - `POST` de recurso aceita `id` gerado pelo cliente (UUIDv7). Reenviar a mesma
      requisição é idempotente: retorna `200` com o recurso já criado em vez de `201`.
    - Listas paginadas usam cursor opaco (`next_cursor`), nunca offset. `next_cursor: null`
      é a última página. O cursor é emitido pelo servidor — não construa um.
  contact:
    name: Meu Auto
  license:
    name: Proprietary
    url: https://meuauto.com.br
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://{host}
    description: Ambiente hospedado
    variables:
      host:
        default: api.meuauto.com.br
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Criar conta
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Conta criada e sessão iniciada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "409":
          description: E-mail já cadastrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Entrar
      description: |
        A resposta de falha é **idêntica** para senha errada e e-mail inexistente, e o
        tempo de resposta também. Distinguir os dois entregaria uma lista de e-mails
        cadastrados.

        Limitado por e-mail (10 tentativas / 15 min) e por IP (60 / 15 min).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Sessão iniciada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      tags: [auth]
      summary: Renovar a sessão
      description: |
        **Rotaciona** o refresh token: o token enviado é revogado e um novo é devolvido.
        Guarde sempre o novo.

        Apresentar um token já rotacionado é tratado como reúso — sinal de que ele foi
        capturado — e **encerra todas as sessões do usuário**, obrigando um novo login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshTokenRequest" }
      responses:
        "200":
          description: Sessão renovada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sair
      description: Idempotente — um token desconhecido também responde `204`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshTokenRequest" }
      responses:
        "204": { description: Sessão encerrada }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Pedir link de redefinição de senha
      description: |
        Responde `202` **sempre**, exista o e-mail ou não. Se o endereço tem conta, um link
        é enviado; o link vale 1 hora e só pode ser usado uma vez.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetRequest" }
      responses:
        "202":
          description: Pedido aceito (resposta idêntica exista ou não a conta)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags: [auth]
      summary: Redefinir a senha
      description: Redefinir **encerra todas as sessões ativas** do usuário.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetConfirm" }
      responses:
        "204": { description: Senha redefinida }
        "401":
          description: Link inválido ou expirado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  # ---------------------------------------------------------------- conta
  /v1/me:
    get:
      operationId: getMe
      tags: [conta]
      summary: Dados da conta
      responses:
        "200":
          description: A conta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateMe
      tags: [conta]
      summary: Atualizar a conta
      description: |
        Só o nome. **Trocar o e-mail não está disponível**: ele é o canal de recuperação da
        conta, e mudá-lo exige verificação no endereço novo e aviso ao antigo.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMeRequest" }
      responses:
        "200":
          description: Conta atualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteMe
      tags: [conta]
      summary: Excluir a conta
      description: |
        **Irreversível.** Apaga a conta e tudo que depende dela — veículos, histórico de
        manutenção, quilometragem, prazos e apólices.

        Exige a senha atual: um token roubado não pode bastar para isso.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteMeRequest" }
      responses:
        "204": { description: Conta e dados excluídos }
        "401":
          description: Senha incorreta ou sessão inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  # ---------------------------------------------------------------- veículos
...
  /v1/vehicles/{vehicleId}/dashboard:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: getDashboard
      tags: [telas]
      summary: A tela principal, em uma requisição
      parameters:
        - name: cost_months
          in: query
          description: Janela do resumo de custos, em meses.
          schema: { type: integer, format: int32, minimum: 1, maximum: 120, default: 12 }
      responses:
        "200":
          description: Veículo, odômetro, contagem de alertas e custos registrados
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dashboard" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/vehicles/{vehicleId}/alerts:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: listAlerts
      tags: [telas]
      summary: Tudo que precisa de ação, de todos os domínios
      description: |
        Manutenção, cuidado recorrente, garantia, IPVA, licenciamento e seguro na mesma
        lista, ordenados por urgência.

        Só entram itens **vencidos** ou **vencendo em breve**. Um plano sem baseline não é
        alerta — é pedido de configuração, e o dashboard o conta separadamente em
        `alerts.needs_baseline`.

        `reference_type` e `reference_id` dizem para onde navegar ao tocar no alerta.
      responses:
        "200":
          description: Os alertas do veículo
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Alert" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/vehicles/{vehicleId}/timeline:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: getTimeline
      tags: [telas]
      summary: Histórico unificado do veículo
      description: |
        Manutenções, leituras manuais de quilometragem e pagamentos de IPVA/licenciamento na
        mesma linha do tempo, mais recente primeiro.

        Leituras de odômetro geradas **por** uma manutenção não aparecem — apareceriam duas
        vezes.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Uma página da linha do tempo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelinePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de acesso obtido em `/v1/auth/login`, `/register` ou `/refresh`. Vale 15
        minutos; renove com o refresh token antes de expirar.
  parameters:
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      description: Tamanho da página. Valores fora da faixa são ajustados, não rejeitados.
      schema: { type: integer, format: int32, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: |
        Cursor opaco devolvido em `next_cursor`. Repasse sem interpretar — o formato é
        interno e pode mudar.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Autenticação necessária, ou sessão inválida/expirada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        Recurso não encontrado.

        Também é a resposta para um recurso que existe mas não pertence ao usuário: dizer
        "você não pode ver isto" confirmaria que ele existe.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationFailed:
      description: |
        Dados inválidos. `details.fields` traz **todos** os campos rejeitados de uma vez,
        com a chave igual à do JSON enviado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: validation_failed
              message: Não foi possível criar a conta.
              details:
                fields:
                  email: Informe um e-mail válido.
                  password: A senha deve ter pelo menos 8 caracteres.
    RateLimited:
      description: Tentativas demais. Aguarde antes de tentar de novo.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Mensagem em pt-BR, exibível ao usuário.
            details:
              type: object
              additionalProperties: true
              description: |
                Dados estruturados específicos do erro. Em `validation_failed` traz
                `fields`; em `odometer_rollback`, o registro vizinho conflitante; em
                `internal`, o `request_id` para suporte.
    ErrorCode:
      type: string
      description: |
        Identificador estável do erro. **É nele que o cliente decide comportamento**, nunca
        na mensagem.

        Valores podem ser acrescentados no futuro — trate um código desconhecido como falha
        genérica, nunca como erro de parsing.
      enum:
        - validation_failed
        - unauthorized
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - odometer_rollback
        - rate_limited
        - internal

    # ------------------------------------------------------------ enums de domínio
...
    AlertKind:
      type: string
      enum: [manutencao, cuidado, garantia, ipva, licenciamento, seguro]
    AlertSeverity:
      type: string
      description: Só dois valores — um alerta é algo que pede ação.
      enum: [vencido, vence_em_breve]
    TimelineEntryKind:
      type: string
      enum: [manutencao, odometro, ipva, licenciamento]

    # ------------------------------------------------------------ auth e conta
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: |
            Mínimo 8 caracteres e nenhuma regra de composição, seguindo o NIST SP 800-63B —
            exigir símbolo e dígito produz "Senha1!" e nada mais seguro.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    PasswordResetConfirm:
      type: object
      required: [token, password]
      properties:
        token: { type: string }
        password: { type: string, minLength: 8, maxLength: 128 }
    UpdateMeRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 120 }
    DeleteMeRequest:
      type: object
      required: [password]
      properties:
        password: { type: string }
    User:
      type: object
      required: [id, name, email, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
    Session:
      type: object
      required: [user, token_type, access_token, expires_at, refresh_token, refresh_expires_at]
      properties:
        user: { $ref: "#/components/schemas/User" }
        token_type: { type: string, enum: [Bearer] }
        access_token: { type: string }
        expires_at: { type: string, format: date-time }
        refresh_token:
          type: string
          description: Guarde com segurança. É rotacionado a cada `/auth/refresh`.
        refresh_expires_at: { type: string, format: date-time }

    # ------------------------------------------------------------ veículo
...
    Alert:
      type: object
      required: [kind, severity, title, reference_type, reference_id]
      properties:
        kind: { $ref: "#/components/schemas/AlertKind" }
        severity: { $ref: "#/components/schemas/AlertSeverity" }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        due_on: { type: string, format: date, nullable: true }
        due_at_km: { type: integer, format: int32, nullable: true }
        remaining_days: { type: integer, format: int32, nullable: true }
        remaining_km: { type: integer, format: int32, nullable: true }
        reference_type:
          type: string
          enum: [maintenance_plan, maintenance_record, obligation, seguro]
          description: Para onde navegar ao tocar no alerta.
        reference_id: { type: string, format: uuid }
    Dashboard:
      type: object
      required: [vehicle, odometer, alerts, costs]
      properties:
        vehicle:
          type: object
          required: [id, brand, model]
          properties:
            id: { type: string, format: uuid }
            brand: { type: string }
            model: { type: string }
            version: { type: string, nullable: true }
            nickname: { type: string, nullable: true }
            plate: { type: string, nullable: true }
        odometer:
          type: object
          required: [current_km]
          properties:
            current_km: { type: integer, format: int32 }
            recorded_on: { type: string, format: date, nullable: true }
        alerts:
          type: object
          required: [overdue, due_soon, needs_baseline, items]
          properties:
            overdue: { type: integer }
            due_soon: { type: integer }
            needs_baseline:
              type: integer
              description: |
                Planos cujo item nunca foi registrado, então não há de onde medir. É pedido
                de configuração, não prazo — por isso é contado aqui e **não** aparece em
                `items` nem em `/alerts`.
            items:
              type: array
              description: Apenas os mais urgentes. A lista completa está em `/alerts`.
              items: { $ref: "#/components/schemas/Alert" }
        costs:
          type: object
          required:
            [period_months, since, maintenance_cents, obligations_cents, seguro_cents, tracked_cents, tracked_categories]
          description: |
            **Não é o custo de rodar o veículo.** Abastecimento e despesas avulsas não
            existem no MVP-1, então `tracked_cents` é a soma apenas do que está listado em
            `tracked_categories`. Rotule na interface como "custo registrado", nunca como
            "custo total".
          properties:
            period_months: { type: integer, format: int32 }
            since: { type: string, format: date }
            maintenance_cents: { type: integer, format: int64 }
            obligations_cents: { type: integer, format: int64 }
            seguro_cents: { type: integer, format: int64 }
            tracked_cents: { type: integer, format: int64 }
            tracked_categories:
              type: array
              items: { type: string }
    TimelineEntry:
      type: object
      required: [kind, id, occurred_on]
      properties:
        kind: { $ref: "#/components/schemas/TimelineEntryKind" }
        id:
          type: string
          format: uuid
          description: Id do recurso de origem — a manutenção, a leitura ou o débito.
        occurred_on: { type: string, format: date }
        title:
          type: string
          nullable: true
          description: |
            Para uma manutenção, os itens do registro separados por vírgula. `null` nos
            demais tipos — a interface monta o rótulo a partir de `kind`.
        subtitle:
          type: string
          nullable: true
          description: |
            Oficina numa manutenção, origem numa leitura de odômetro, ano de referência num
            débito.
        amount_cents:
          type: integer
          format: int64
          nullable: true
          description: '`null` numa leitura de odômetro — não envolve dinheiro.'
        mileage_km:
          type: integer
          format: int32
          nullable: true
          description: '`null` num pagamento de tributo — não envolve quilometragem.'
    TimelinePage:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/TimelineEntry" }
        next_cursor: { type: string, nullable: true }